third_party/blink/web_tests/external/wpt/css/css-fonts/variations/font-shorthand-expected.txt

This is a testharness.js-based test.
[FAIL] Font shorthand: 'oblique' without slant angle
  assert_equals: Font shorthand expected style: 'oblique' without slant angle expected "oblique" but got "italic"
[FAIL] Font shorthand: 'oblique' followed by valid small weight
  assert_equals: Font shorthand expected style: 'oblique' followed by valid small weight expected "oblique" but got "italic"
[FAIL] Font shorthand: 'oblique' followed by valid large weight
  assert_equals: Font shorthand expected style: 'oblique' followed by valid large weight expected "oblique" but got "italic"
[FAIL] Font shorthand: 'oblique' followed by valid calc() weight
  assert_equals: Font shorthand expected style: 'oblique' followed by valid calc() weight expected "oblique" but got "italic"
[FAIL] Font shorthand: 'oblique' followed by a to-be-clamped calc() weight
  assert_equals: Font shorthand expected style: 'oblique' followed by a to-be-clamped calc() weight expected "oblique" but got "italic"
[FAIL] Font shorthand: calc() weight folowed by 'oblique'
  assert_equals: Font shorthand expected style: calc() weight folowed by 'oblique' expected "oblique" but got "italic"
Harness: the test ran to completion.
